refactor: update directories of widgets/
