+++ src/services/dashboards/widgets/_base/base-pie/widget-config.ts
@@ -1,6 +1,6 @@
-import type { WidgetConfig } from '@/services/dashboards/widgets/config';
-import { GROUP_BY } from '@/services/dashboards/widgets/config';
-import { GROUP_BY_ITEM_MAP } from '@/services/dashboards/widgets/view-config';
+import type { WidgetConfig } from '@/services/dashboards/widgets/_configs/config';
+import { GROUP_BY } from '@/services/dashboards/widgets/_configs/config';
+import { GROUP_BY_ITEM_MAP } from '@/services/dashboards/widgets/_configs/view-config';
 
 const basePieWidgetConfig: WidgetConfig = {
     widget_config_id: 'basePie',
