+++ src/services/dashboards/widgets/_base/base-trend/widget-config.ts
@@ -1,4 +1,4 @@
-import type { WidgetConfig } from '@/services/dashboards/widgets/config';
+import type { WidgetConfig } from '@/services/dashboards/widgets/_configs/config';
 
 const baseTrendWidgetConfig: WidgetConfig = {
     widget_config_id: 'baseTrend',
