+++ src/services/dashboards/widgets/aws-data-transfer-cost-trend/widget-config.ts
@@ -1,6 +1,6 @@
-import type { WidgetConfig } from '@/services/dashboards/widgets/config';
-import { CHART_TYPE, GRANULARITY } from '@/services/dashboards/widgets/config';
-import { excludePropertiesFromDashboardCommonWidgetConfig } from '@/services/dashboards/widgets/widget-options-schema-helper';
+import type { WidgetConfig } from '@/services/dashboards/widgets/_configs/config';
+import { CHART_TYPE, GRANULARITY } from '@/services/dashboards/widgets/_configs/config';
+import { excludePropertiesFromDashboardCommonWidgetConfig } from '@/services/dashboards/widgets/_helpers/widget-options-schema-helper';
 
 const {
     default_properties,
