+++ src/services/dashboards/widgets/cost-donut/widget-config.ts
@@ -1,4 +1,4 @@
-import type { WidgetConfig } from '@/services/dashboards/widgets/config';
+import type { WidgetConfig } from '@/services/dashboards/widgets/_configs/config';
 
 const costDonutWidgetConfig: Partial<WidgetConfig> = {
     widget_config_id: 'costDonut',
