+++ src/services/dashboards/widgets/cost-trend/widget-config.ts
@@ -1,6 +1,6 @@
-import type { WidgetConfig } from '@/services/dashboards/widgets/config';
-import { CHART_TYPE, GROUP_BY } from '@/services/dashboards/widgets/config';
-import { GROUP_BY_ITEM_MAP } from '@/services/dashboards/widgets/view-config';
+import type { WidgetConfig } from '@/services/dashboards/widgets/_configs/config';
+import { CHART_TYPE, GROUP_BY } from '@/services/dashboards/widgets/_configs/config';
+import { GROUP_BY_ITEM_MAP } from '@/services/dashboards/widgets/_configs/view-config';
 
 const costTrendWidgetConfig: Partial<WidgetConfig> = {
     widget_config_id: 'costTrend',
