+++ src/services/dashboards/widgets/monthly-cost/widget-config.ts
@@ -1,5 +1,5 @@
-import type { WidgetConfig } from '@/services/dashboards/widgets/config';
-import { GRANULARITY } from '@/services/dashboards/widgets/config';
+import type { WidgetConfig } from '@/services/dashboards/widgets/_configs/config';
+import { GRANULARITY } from '@/services/dashboards/widgets/_configs/config';
 
 const monthlyCostWidgetConfig: WidgetConfig = {
     widget_config_id: 'monthlyCost',
